moved remaining methods to service provider
